backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
sentence-transformers==2.2.2
numpy==1.24.3
python-dotenv==1.0.0
google-generativeai==0.3.1
scikit-learn==1.3.0
